backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
pydantic-settings==2.6.1
python-dotenv==1.2.2
SQLAlchemy==2.0.48
psycopg2-binary==2.9.10
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.7
email-validator==2.2.0
bcrypt==4.2.1
cryptography==44.0.0
httpx==0.27.0
requests==2.32.3
chromadb==0.5.3
sentence-transformers==3.0.1
langchain-text-splitters==0.2.2
pypdf==4.2.0
supabase==2.5.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
annotated-types==0.7.0
typing_extensions==4.15.0


